requirements.txt
streamlit>=1.28.0
flask>=2.3.0
flask-cors>=4.0.0
langchain>=0.0.300
langchain-openai>=0.0.2
azure-search-documents>=11.4.0
azure-storage-blob>=12.19.0
azure-cosmos>=4.5.0
openai>=1.3.0
python-dotenv>=1.0.0
PyPDF2>=3.0.0
pdfplumber>=0.9.0
requests>=2.31.0
pandas>=2.0.0
numpy>=1.26.0
tiktoken>=0.5.0
